% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_data_download.R
\name{dataDownloadLink}
\alias{dataDownloadLink}
\title{Data Download Link}
\usage{
dataDownloadLink(
  id,
  label = "Save",
  tooltip = "Save the data in various formats"
)
}
\arguments{
\item{id}{module id}

\item{label}{Label for the download link}

\item{tooltip}{tooltip text shown on hover}
}
\description{
Data Download Link
}
