% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_data_download.R
\name{dataDownloadServer}
\alias{dataDownloadServer}
\title{Data Download Server}
\usage{
dataDownloadServer(input, output, session, data_func, filename_func)
}
\arguments{
\item{data_func}{reactive function providing the data}

\item{filename_func}{reactive function returning the default file name}
}
\description{
Data Download Server
}
