% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{micrologger-package}
\alias{micrologger}
\alias{micrologger-package}
\title{micrologger: GUI for GEOM µLoggers}
\description{
Provides a graphical user interface to interact with µLoggers including setting up experiments, visualizating data, and sending commands.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KopfLab/micrologger}
  \item Report bugs at \url{https://github.com/KopfLab/micrologger/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{seb.kopf@gmail.com}

}
\keyword{internal}
