% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_write.R
\name{ml_archive_experiment}
\alias{ml_archive_experiment}
\title{Archive an experiment}
\usage{
ml_archive_experiment(exp_id, con = db())
}
\arguments{
\item{exp_id}{id of the experiment to archive (single exp_id)}

\item{con}{database connection (defaults to \code{db()})}
}
\description{
Marks the experiment as archived and releases all devices it controlled.
}
