% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_write.R
\name{ml_experiment_stop_recording}
\alias{ml_experiment_stop_recording}
\title{Stop recording an experiment}
\usage{
ml_experiment_stop_recording(exp_id, con = db())
}
\arguments{
\item{exp_id}{id of the experiment to stop recording (single exp_id)}

\item{con}{database connection (defaults to \code{db()})}
}
\description{
Marks the experiment as no longer recording (pauses recording).
}
