% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_read.R
\name{ml_get_experiments}
\alias{ml_get_experiments}
\title{Retrieve experiments}
\usage{
ml_get_experiments(
  group_id = NULL,
  filter = NULL,
  convert_to_TZ = Sys.timezone(),
  con = db()
)
}
\arguments{
\item{group_id}{experiments from which group to fetch}

\item{filter}{optional expression to filter the experiments}

\item{convert_to_TZ}{timezone to convert last_recording_change to (skipped if NULL)}

\item{con}{database connection (defaults to db())}
}
\value{
experiments
}
\description{
Retrieve experiments
}
