% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_read.R
\name{ml_get_groups}
\alias{ml_get_groups}
\title{Get user groups}
\usage{
ml_get_groups(group_id = NULL, con = db())
}
\arguments{
\item{group_id}{which group(s) to fetch (all if NULL)}

\item{con}{database connection (defaults to db())}
}
\description{
Get user groups
}
