% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_read.R
\name{ml_get_logs}
\alias{ml_get_logs}
\title{read data logs}
\usage{
ml_get_logs(
  exp_id,
  filter = NULL,
  select = everything(),
  max_rows = NULL,
  cache = TRUE,
  include_snapshots = TRUE,
  con = db(),
  parse = TRUE,
  timezone = Sys.timezone(),
  experiments = NULL
)
}
\arguments{
\item{exp_id}{which experiment id(s) to fetch logs for}

\item{filter}{optional expression to filter the logs (applied after caching)}

\item{select}{tidyselect columns to return from the query}

\item{max_rows}{optional cap on the number of rows retrieved}

\item{cache}{whether to cache raw logs to disk and reuse the cache}

\item{include_snapshots}{whether to also include device snapshots}

\item{con}{database connection (defaults to db())}

\item{parse}{whether to parse the database logs}

\item{timezone}{what timezone to use for parsing}

\item{experiments}{tibble with experiment information if this should be merged in}
}
\description{
read data logs
}
