% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run.R
\name{ml_run_gui}
\alias{ml_run_gui}
\title{Run the user interface}
\usage{
ml_run_gui(
  user_id,
  user_groups,
  user_is_admin = FALSE,
  user_first_name = NA_character_,
  user_last_name = NA_character_,
  token = keyring::key_get("particle"),
  timezone = Sys.timezone(),
  options = list(),
  uiPattern = "/",
  enableBookmarking = "url"
)
}
\arguments{
\item{token}{particle access token (retrieved from keyring by default), save a token in your keyring with particle_store_token()}

\item{timezone}{the timezone to user for datetime calculations}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}
}
\description{
Set database credentials ahead of time via \link{ml_set_connection_credentials}.
}
