% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_utils.R
\name{ml_set_connection_credentials}
\alias{ml_set_connection_credentials}
\title{Set DB connection credentials}
\usage{
ml_set_connection_credentials(
  host = keyring::key_get("MICROLOGGER_DB_HOST"),
  user = keyring::key_get("MICROLOGGER_DB_USER"),
  password = keyring::key_get("MICROLOGGER_DB_PASSWORD")
)
}
\description{
Set DB connection credentials
}
