% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_utils.R
\name{ml_store_connection_credentials_user}
\alias{ml_store_connection_credentials_user}
\title{Securley store the access information using your OS keyring
Call \link{ml_set_connection_credentials} for a new R session to transfer these to environmental variables.}
\usage{
ml_store_connection_credentials_user()
}
\description{
Securley store the access information using your OS keyring
Call \link{ml_set_connection_credentials} for a new R session to transfer these to environmental variables.
}
