% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_logs.R
\name{ml_summarize_log_devices}
\alias{ml_summarize_log_devices}
\title{summarize log devices}
\usage{
ml_summarize_log_devices(
  logs,
  include_snapshot_only_values = FALSE,
  include_text_values = FALSE
)
}
\description{
summarize log devices
}
