% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_write.R
\name{ml_unlink_device_from_experiment}
\alias{ml_unlink_device_from_experiment}
\title{Remove experiment devices}
\usage{
ml_unlink_device_from_experiment(exp_id, core_id, con = db())
}
\arguments{
\item{exp_id}{id of the experiment to unlink devices from (single exp_id)}

\item{core_id}{can be multiple ids}

\item{con}{database connection (defaults to \code{db()})}
}
\description{
Remove experiment devices
}
