% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_write.R
\name{ml_unlink_devices_from_experiment}
\alias{ml_unlink_devices_from_experiment}
\title{Remove experiment devices}
\usage{
ml_unlink_devices_from_experiment(exp_id, core_ids, con = db())
}
\description{
Remove experiment devices
}
