% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_write.R
\name{ml_update_experiment_info}
\alias{ml_update_experiment_info}
\title{Update experiment information}
\usage{
ml_update_experiment_info(
  exp_id,
  name = NULL,
  description = NULL,
  notes = NULL,
  con = db()
)
}
\arguments{
\item{exp_id}{id of the experiment to update (single id)}

\item{name}{new experiment name, if provided}

\item{description}{new experiment description, if provided}

\item{notes}{new experiment notes, if provided}

\item{con}{database connection (defaults to \code{db()})}
}
\value{
whether the update was successful or not
}
\description{
Update experiment information
}
