% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_plot_download.R
\name{plotDownloadLink}
\alias{plotDownloadLink}
\title{Plot Download Link}
\usage{
plotDownloadLink(id, label = "Save", tooltip = "Save the plot as a PDF")
}
\arguments{
\item{id}{module id}

\item{label}{Label for the download link}

\item{tooltip}{tooltip text shown on hover}
}
\description{
Plot Download Link
}
