% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_write.R
\name{ml_add_experiment}
\alias{ml_add_experiment}
\title{Add new experiment}
\usage{
ml_add_experiment(
  group_id,
  user_id,
  user_first = NA_character_,
  user_last = NA_character_,
  con = db()
)
}
\description{
Add new experiment
}
