% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_write.R
\name{ml_claim_device_for_experiment}
\alias{ml_claim_device_for_experiment}
\title{Claim device for a paritcular experiment}
\usage{
ml_claim_device_for_experiment(exp_id, core_id, con = db())
}
\description{
Claim device for a paritcular experiment
}
