% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_read.R
\name{ml_get_devices}
\alias{ml_get_devices}
\title{Retrieve devices including their structures and who is currently using them}
\usage{
ml_get_devices(
  group_id = NULL,
  filter = NULL,
  select = everything(),
  con = db()
)
}
\arguments{
\item{group_id}{devices from which group(s) to fetch}
}
\value{
devices
}
\description{
Retrieve devices including their structures and who is currently using them
}
