% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_read.R
\name{ml_get_experiment_devices}
\alias{ml_get_experiment_devices}
\title{Retrieve experiment devices}
\usage{
ml_get_experiment_devices(
  exp_id = NULL,
  filter = NULL,
  select = everything(),
  con = db()
)
}
\value{
experiments_devices
}
\description{
Returns experiment-devices merged with devices table and structure information
}
