% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_read.R
\name{ml_get_logs}
\alias{ml_get_logs}
\title{read data logs}
\usage{
ml_get_logs(
  exp_id,
  filter = NULL,
  select = everything(),
  max_rows = NULL,
  cache = TRUE,
  include_snapshots = TRUE,
  con = db(),
  parse = TRUE,
  timezone = Sys.timezone(),
  experiments = NULL
)
}
\arguments{
\item{parse}{whether to parse the database logs}

\item{timezone}{what timezone to use for parsing}

\item{experiments}{tibble with experiment information if this should be merged in}
}
\description{
read data logs
}
