% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_write.R
\name{ml_link_devices_to_experiment}
\alias{ml_link_devices_to_experiment}
\title{Add/update for the experiment_devices table}
\usage{
ml_link_devices_to_experiment(
  exp_id,
  core_ids,
  labels = NA_character_,
  experiment_devices = tibble(exp_id, core_id = core_ids, label = labels),
  con = db()
)
}
\description{
Add/update for the experiment_devices table
}
