% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_write.R
\name{ml_release_device_from_experiment}
\alias{ml_release_device_from_experiment}
\title{Free device from an experiment so it can be added to others}
\usage{
ml_release_device_from_experiment(exp_id, core_id, con = db())
}
\description{
Free device from an experiment so it can be added to others
}
