% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_db_write.R
\name{ml_unlink_device_from_experiment}
\alias{ml_unlink_device_from_experiment}
\title{Remove experiment devices}
\usage{
ml_unlink_device_from_experiment(exp_id, core_id, con = db())
}
\arguments{
\item{core_id}{can be multiple ids}
}
\description{
Remove experiment devices
}
